+++ src/actions/types.js
@@ -1,5 +1,10 @@
+//TEXTAREA and MDPREVIEW
 export const ADD_TEXT = "ADD_TEXT";
 export const PARSE_MD = "PARSE_MD";
 
+//GOOGLE AUTH
 export const SIGN_IN = "SIGN_IN";
 export const SIGN_OUT = "SIGN_OUT";
+
+//FORM PAGE
+export const ADD_FORM_VALUES = "ADD_FORM_VALUES";
