+++ src/components/App.js
@@ -1,8 +1,7 @@
 import React from "react";
-import FormComponent from "./YAMLComponents/FormComponent";
+import FormComponent from "./lessonForm/FormComponent";
 import EndPage from "./frontPageComponents/EndPage";
-import MarkdownEditor from "./mdPreviewComponents/MarkdownEditor";
-// import Editor from "./mdPreviewComponents/Editor";
+import Editor from "./mdTextEditor/editor/Editor";
 import IndexPage from "./frontPageComponents/IndexPage";
 import Overview from "./mypage/Overview";
 import { BrowserRouter, Route } from "react-router-dom";
@@ -13,9 +12,9 @@ const App = () => {
       <BrowserRouter>
         <div>
           <Route exact path="/" component={IndexPage} />
-          <Route path="/myPage" component={Overview} />
-          <Route path="/createNewLesson" component={FormComponent} />
-          <Route path="/editor" component={MarkdownEditor} />
+          <Route exact path="/myPage" component={Overview} />
+          <Route exact path="/createNewLesson" component={FormComponent} />
+          <Route exact path="/editor" component={Editor} />
           <Route exact path="/endpage" component={EndPage} />
         </div>
       </BrowserRouter>
