+++ src/components/frontPageComponents/IndexPage.js
@@ -3,6 +3,7 @@ import "./style.css";
 import { connect } from "react-redux";
 import { Link } from "react-router-dom";
 import GoogleAuth from "../GoogleAuth";
+import ProfileMenu from "../ProfileMenu";
 
 const IndexPage = props => {
   const indPage = () => {
@@ -11,10 +12,10 @@ const IndexPage = props => {
         <div className="StartPageContainer">
           <div className="ui right aligned grid">
             <div className="right floated column">
-              <GoogleAuth />
+              <ProfileMenu />
             </div>
           </div>
-
+          <div style={{ margin: "1rem" }} />
           <div className="quote">
             <p>
               "Programs must be written for people to read, and only
@@ -22,15 +23,8 @@ const IndexPage = props => {
             </p>
             <p id="author">Harold Abelson </p>
           </div>
-          <div style={{ margin: "45px" }} />
-          <div
-            style={{
-              margin: "20, 20, 5, 20",
-              textAlign: "center"
-            }}
-          >
-            <h2>Velkommen {props.firstName}</h2>
-          </div>
+          <div />
+
           <div className="btnDiv">
             <Link to="/myPage" className="link">
               <button className="btn">Min side</button>
@@ -54,18 +48,8 @@ const IndexPage = props => {
             </p>
             <p id="author">Harold Abelson </p>
           </div>
-          <div style={{}} />
-          <div
-            style={{
-              marginTop: "10vh",
-              marginBottom: "2vh",
-              alignItems: "center",
-              textAlign: "center"
-            }}
-            className="ui container"
-          >
-            <h2>Logg inn:</h2>
-          </div>
+          <div />
+          <div className="ui center container" />
           <div className="center">
             <GoogleAuth />
           </div>
