+++ src/components/lessonForm/Page3.js
@@ -48,64 +48,60 @@ for (let i = 0; i < topicLen; i++) {
 
 const Page3 = props => {
   return (
-    <div className="">
-      <div className="ui grid">
-        <div className="sixteen wide column">
-          <div className="field">
-            <label>
-              <h3>
-                {FORM_TEXT.TITLE.heading}
-                <span className="test"> (nødvendig)</span>
-              </h3>
-            </label>
-            <input
-              autoFocus
-              autoComplete="off"
-              type="text"
-              name="title"
-              placeholder={FORM_TEXT.TITLE.placeholder}
-              value={props.state.title}
-              onChange={props.changeHandler}
-            />
-          </div>
-        </div>
-        <div className="validateError">{props.state.titleErr}</div>
-      </div>
+    <React.Fragment>
+      <div className="ui form">
+        <div className="field">
+          <label>
+            <h3 className="formLabel">
+              {FORM_TEXT.TITLE.heading}
+              <span className="requiredText"> (obligatorisk)</span>
+            </h3>
+          </label>
+          <input
+            autoFocus
+            autoComplete="off"
+            type="text"
+            name="title"
+            placeholder={FORM_TEXT.TITLE.placeholder}
+            value={props.state.title}
+            onChange={props.changeHandler}
+          />
 
-      <div id="bigScreen" className="ui grid">
-        <div className="eight wide column">
+          <div className="validateError">{props.state.titleErr}</div>
+        </div>
+        <div id="bigScreen" className="two fields">
           <div className="field">
             <label>
-              <h3>{YML_TEXT.subject}</h3>
+              <h3 className="formLabel">{YML_TEXT.subject}</h3>
             </label>
-            <div className="ui segment">
-              <div className="ui stackable two column grid">
+            <div id="subjectCheckbox" className="ui segment">
+              <div className="ui grid">
                 <YMLtagsSubject checkboxHandler={props.checkboxHandler} />
               </div>
             </div>
           </div>
-        </div>
-        <div className="eight wide column">
+
           <div className="field">
             <label>
-              <h3>{YML_TEXT.topic}</h3>
+              <h3 className="formLabel">{YML_TEXT.topic}</h3>
             </label>
-            <div className="ui segment">
-              <div className="ui stackable two column grid">
-                <YMLtagsTopic checkboxHandler={props.checkboxHandler} />
+            <div id="topicCheckbox" className="ui segment">
+              <div className="ui grid">
+                <div className="four wide column">
+                  <YMLtagsTopic checkboxHandler={props.checkboxHandler} />
+                </div>
               </div>
             </div>
           </div>
         </div>
-      </div>
-      <div style={{ margin: "30px" }} />
-      <div id="smallScreen" className="ui grid">
-        <div className="sixteen wide column">
-          <div className="field">
+
+        {/* Only if smallScreen */}
+        <div id="smallScreen" className="ui container">
+          <div id="smallSubjectField" className="field">
             <label>
-              <h3>{YML_TEXT.subject}</h3>
+              <h3 className="formLabel">{YML_TEXT.subject}</h3>
             </label>
-            <div style={{ marginRight: "-30px" }}>
+            <div>
               <Dropdown
                 placeholder="Velg fag"
                 fluid
@@ -117,14 +113,12 @@ const Page3 = props => {
               />
             </div>
           </div>
-        </div>
-        <div style={{ margin: "30px" }} />
-        <div className="sixteen wide column">
-          <div className="field">
+
+          <div id="smallTopicField" className="field">
             <label>
-              <h3>{YML_TEXT.topic}</h3>
+              <h3 className="formLabel">{YML_TEXT.topic}</h3>
             </label>
-            <div style={{ marginLeft: "-13px", marginRight: "-15px" }}>
+            <div>
               <Dropdown
                 placeholder="Velg tema"
                 fluid
@@ -137,9 +131,8 @@ const Page3 = props => {
             </div>
           </div>
         </div>
-        <div style={{ margin: "100px" }} />
       </div>
-    </div>
+    </React.Fragment>
   );
 };
