--- src/components/mdPreviewComponents/largeScreen/ControlPanel.js
@@ -1,158 +1,0 @@
-import React from "react";
-import { connect } from "react-redux";
-import Buttons from "../Buttons";
-import ProfileMenu from "../../ProfileMenu";
-import {
-  emphasis,
-  undoRedo,
-  saveLoadNew,
-  image,
-  lists,
-  sections,
-  code
-} from "../settingsFiles/buttonConfig.js";
-
-class ControlPanel extends React.Component {
-  handleButtonClick = (
-    bTitle,
-    output,
-    cursorIntON,
-    cursorIntOFF,
-    endOutput
-  ) => {
-    this.props.handleButtonClick(
-      bTitle,
-      output,
-      cursorIntON,
-      cursorIntOFF,
-      endOutput
-    );
-  };
-
-  render() {
-    return (
-      <div className="ui segment grid buttonBorder">
-        <div className="ui icon buttons">
-          {emphasis.map(element => (
-            <Buttons
-              bTitle={element.bTitle}
-              icon={element.icon}
-              output={element.output}
-              title={element.title}
-              cursorIntON={element.cursorIntON}
-              cursorIntOFF={element.cursorIntOFF}
-              endOutput={element.endOutput}
-              onButtonClick={this.handleButtonClick}
-              shortcutKey={element.shortcut}
-            />
-          ))}
-        </div>
-
-        <div className="ui icon buttons">
-          {undoRedo.map(element => (
-            <Buttons
-              bTitle={element.bTitle}
-              icon={element.icon}
-              output={element.output}
-              title={element.title}
-              cursorIntON={element.cursorIntON}
-              cursorIntOFF={element.cursorIntOFF}
-              endOutput={element.endOutput}
-              onButtonClick={this.handleButtonClick}
-              shortcutKey={element.shortcut}
-            />
-          ))}
-        </div>
-
-        <div className="ui icon buttons">
-          {saveLoadNew.map(element => (
-            <Buttons
-              bTitle={element.bTitle}
-              icon={element.icon}
-              output={element.output}
-              title={element.title}
-              cursorIntON={element.cursorIntON}
-              cursorIntOFF={element.cursorIntOFF}
-              endOutput={element.endOutput}
-              onButtonClick={this.handleButtonClick}
-              shortcutKey={element.shortcut}
-            />
-          ))}
-        </div>
-
-        <div className="ui icon buttons">
-          {image.map(element => (
-            <Buttons
-              bTitle={element.bTitle}
-              icon={element.icon}
-              output={element.output}
-              title={element.title}
-              cursorIntON={element.cursorIntON}
-              cursorIntOFF={element.cursorIntOFF}
-              endOutput={element.endOutput}
-              onButtonClick={this.handleButtonClick}
-              shortcutKey={element.shortcut}
-            />
-          ))}
-        </div>
-
-        <div className="ui icon buttons">
-          {lists.map(element => (
-            <Buttons
-              bTitle={element.bTitle}
-              icon={element.icon}
-              output={element.output}
-              title={element.title}
-              cursorIntON={element.cursorIntON}
-              cursorIntOFF={element.cursorIntOFF}
-              endOutput={element.endOutput}
-              onButtonClick={this.handleButtonClick}
-              shortcutKey={element.shortcut}
-            />
-          ))}
-        </div>
-        <div className="right floated two wide column">
-          <ProfileMenu />
-        </div>
-        <div id="sections" className="ui icon buttons">
-          {sections.map(element => (
-            <Buttons
-              bTitle={element.bTitle}
-              icon={element.icon}
-              output={element.output}
-              title={element.title}
-              cursorIntON={element.cursorIntON}
-              cursorIntOFF={element.cursorIntOFF}
-              endOutput={element.endOutput}
-              onButtonClick={this.handleButtonClick}
-              shortcutKey={element.shortcut}
-            />
-          ))}
-        </div>
-        <div id="code" className="ui icon buttons">
-          {code.map(element => (
-            <Buttons
-              bTitle={element.bTitle}
-              icon={element.icon}
-              output={element.output}
-              title={element.title}
-              cursorIntON={element.cursorIntON}
-              cursorIntOFF={element.cursorIntOFF}
-              endOutput={element.endOutput}
-              onButtonClick={this.handleButtonClick}
-              shortcutKey={element.shortcut}
-            />
-          ))}
-        </div>
-      </div>
-    );
-  }
-}
-
-const mapStateToProps = state => {
-  return {
-    imageUrl: state.auth.imageUrl
-  };
-};
-
-export default connect(mapStateToProps)(ControlPanel);
