+++ src/components/mdTextEditor/editor/Editor.js
@@ -1,17 +1,15 @@
-import "./smallscreen.css";
+import "./editor.css";
 import React from "react";
-import { Redirect } from "react-router-dom";
 import { connect } from "react-redux";
+import { Redirect } from "react-router-dom";
+import { Button, Icon, Popup } from "semantic-ui-react";
 import { addText, parseMD } from "../../../actions";
 import MDTextArea from "./MDTextArea";
-import MDPreview from "./MDPreview";
+import MDPreview from "../mdPreview/MDPreview";
 import { mdParser } from "../../../utils/mdParser";
-import ControlPanel from "./ControlPanel";
-import ControlPanel2 from "./ControlPanel2";
-import ControlPanelPreview from "./ControlPanelPreview";
-import PageButtons from "../../PageButtons";
+import ControlPanel from "./controlpanel/ControlPanel";
 import ImagePopup from "../ImagePopup";
-import GoogleAuth from "../../GoogleAuth";
+import PageButtons from "../../PageButtons";
 import {
   SAVING,
   SAVED,
@@ -19,6 +17,17 @@ import {
   PHOTO_TEXT,
   NAV_BUTTONS
 } from "../settingsFiles/languages/editor_NO";
+import {
+  SHORTCUTKEY,
+  KEY_COMBINATIONS as KEY,
+  emphasis,
+  undoRedo,
+  saveLoadNew,
+  image,
+  lists,
+  sections,
+  code
+} from "../settingsFiles/buttonConfig";
 
 // check if buttons is pressed
 var isButtonOn = {
@@ -46,6 +55,12 @@ var isButtonOn = {
   inline: true,
   codeblock: true
 };
+
+// dynamic list with all the keyboard shortcut chars from ./settingFiles/buttonConfig.js
+var shortcutKeys = [];
+for (let i = 0; i < Object.values(KEY).length; i++) {
+  shortcutKeys.push(Object.values(KEY)[i][Object.values(KEY)[i].length - 1]);
+}
 
 // Count input char for automatic newline at 80 chars
 var charCounter = 0;
@@ -82,12 +97,11 @@ class Editor extends React.Component {
     super(props);
 
     this.state = {
-      test: true,
-      showPreview: false,
+      preview: false,
+      isEditor: true,
+      editorRedirect: "",
       images: [],
       counter: 0,
-      textValue: "",
-      mdValue: "",
       buttonValues: isButtonOn,
       redirect: null
     };
@@ -110,12 +124,6 @@ class Editor extends React.Component {
 
   // auto save after a couple of seconds
   componentDidUpdate() {
-    if (
-      window.innerWidth > 600 ||
-      window.innerHeight / window.innerWidth < 1.4
-    ) {
-      this.props.update();
-    }
     if (this.state.counter === 4 && this.props.mdText.length > 0) {
       autoSaveMessage = SAVED;
     } else if (this.state.counter === 0) {
@@ -126,7 +134,7 @@ class Editor extends React.Component {
 
   render() {
     // Submithandler
-    const mySubmitHandler = event => {
+    const submitHandler = event => {
       event.preventDefault();
 
       console.log("text submitted");
@@ -238,6 +246,23 @@ class Editor extends React.Component {
     const onTextareaKeyDown = event => {
       cursorPositionStart = event.target.selectionStart;
       cursorPositionEnd = event.target.selectionEnd;
+
+      // prevents default value on shortcut keys
+      if (
+        (event.ctrlKey && SHORTCUTKEY === "ctrl") ||
+        (event.altKey && SHORTCUTKEY === "alt") ||
+        (event.metaKey && SHORTCUTKEY === "command") ||
+        (event.shiftKey && SHORTCUTKEY === "shift")
+      ) {
+        if (shortcutKeys.includes(event.key)) {
+          event.preventDefault();
+        }
+      }
+
+      // if spacebar, update undo
+      if (event.key === " ") {
+        setUndo();
+      }
 
       // if input is enter, update undo and do list functions if list.
       if (event.key === "Enter") {
@@ -702,61 +727,308 @@ class Editor extends React.Component {
       }
     };
 
-    const returnPreview = () => {
-      if (!this.state.showPreview) {
-        this.setState({ showPreview: true });
-      } else {
-        this.setState({ showPreview: false });
+    // Make keyboard shortcuts with React Hotkeys
+    // config in ./settingsFiles/buttonConfig.js
+    const keyMap = {
+      BOLD: KEY.bold.join(""),
+      ITALIC: KEY.italic.join(""),
+      HEADING: KEY.heading.join(""),
+      STRIKETHROUGH: KEY.strikethrough.join(""),
+      UNDO: KEY.undo.join(""),
+      REDO: KEY.redo.join(""),
+      NEW: KEY.new.join(""),
+      LOAD: KEY.load.join(""),
+      SAVE: KEY.save.join(""),
+      IMAGE: KEY.image.join(""),
+      LISTUL: KEY.listul.join(""),
+      LISTOL: KEY.listol.join(""),
+      CHECKLIST: KEY.listcheck.join(""),
+      ACTIVITY: KEY.activity.join(""),
+      INTRO: KEY.intro.join(""),
+      CHECK: KEY.check.join(""),
+      TIP: KEY.tip.join(""),
+      PROTIP: KEY.protip.join(""),
+      CHALLENGE: KEY.challenge.join(""),
+      FLAG: KEY.flag.join(""),
+      TRY: KEY.try.join(""),
+      INLINE: KEY.inline.join(""),
+      CODEBLOCK: KEY.codeblock.join("")
+    };
+
+    // keyboard shortcut actions.  Settings in ./settingsFiles/buttonConfig.js
+    // needs to be updated manually with same parameters as buttonConfig
+    // SORRY FOR WET CODE AND MANUAL UPDATE. NEED HELP FIXING THIS
+    const handlers = {
+      // emphasis
+      BOLD: () =>
+        handleButtonClick(
+          emphasis[0].bTitle,
+          emphasis[0].output,
+          emphasis[0].cursorIntON,
+          emphasis[0].cursorIntOFF,
+          emphasis[0].endOutput
+        ),
+      ITALIC: () =>
+        handleButtonClick(
+          emphasis[1].bTitle,
+          emphasis[1].output,
+          emphasis[1].cursorIntON,
+          emphasis[1].cursorIntOFF,
+          emphasis[1].endOutput
+        ),
+      HEADING: () =>
+        handleButtonClick(
+          emphasis[2].bTitle,
+          emphasis[2].output,
+          emphasis[2].cursorIntON,
+          emphasis[2].cursorIntOFF,
+          emphasis[2].endOutput
+        ),
+      STRIKETHROUGH: () =>
+        handleButtonClick(
+          emphasis[3].bTitle,
+          emphasis[3].output,
+          emphasis[3].cursorIntON,
+          emphasis[3].cursorIntOFF,
+          emphasis[3].endOutput
+        ),
+
+      //undoRedo
+      UNDO: () =>
+        handleButtonClick(
+          undoRedo[0].bTitle,
+          undoRedo[0].output,
+          undoRedo[0].cursorIntON,
+          undoRedo[0].cursorIntOFF,
+          undoRedo[0].endOutput
+        ),
+      REDO: () =>
+        handleButtonClick(
+          undoRedo[1].bTitle,
+          undoRedo[1].output,
+          undoRedo[1].cursorIntON,
+          undoRedo[1].cursorIntOFF,
+          undoRedo[1].endOutput
+        ),
+
+      //saveLoadNew
+      NEW: () =>
+        handleButtonClick(
+          saveLoadNew[0].bTitle,
+          saveLoadNew[0].output,
+          saveLoadNew[0].cursorIntON,
+          saveLoadNew[0].cursorIntOFF,
+          saveLoadNew[0].endOutput
+        ),
+      LOAD: () =>
+        handleButtonClick(
+          saveLoadNew[1].bTitle,
+          saveLoadNew[1].output,
+          saveLoadNew[1].cursorIntON,
+          saveLoadNew[1].cursorIntOFF,
+          saveLoadNew[1].endOutput
+        ),
+      SAVE: () =>
+        handleButtonClick(
+          saveLoadNew[2].bTitle,
+          saveLoadNew[2].output,
+          saveLoadNew[2].cursorIntON,
+          saveLoadNew[2].cursorIntOFF,
+          saveLoadNew[2].endOutput
+        ),
+
+      //image
+      IMAGE: () =>
+        handleButtonClick(
+          image[0].bTitle,
+          image[0].output,
+          image[0].cursorIntON,
+          image[0].cursorIntOFF,
+          image[0].endOutput
+        ),
+
+      //lists
+      LISTUL: () =>
+        handleButtonClick(
+          lists[0].bTitle,
+          lists[0].output,
+          lists[0].cursorIntON,
+          lists[0].cursorIntOFF,
+          lists[0].endOutput
+        ),
+      LISTOL: () =>
+        handleButtonClick(
+          lists[1].bTitle,
+          lists[1].output,
+          lists[1].cursorIntON,
+          lists[1].cursorIntOFF,
+          lists[1].endOutput
+        ),
+      CHECKLIST: () =>
+        handleButtonClick(
+          lists[2].bTitle,
+          lists[2].output,
+          lists[2].cursorIntON,
+          lists[2].cursorIntOFF,
+          lists[2].endOutput
+        ),
+
+      //sections
+      ACTIVITY: () =>
+        handleButtonClick(
+          sections[0].bTitle,
+          sections[0].output,
+          sections[0].cursorIntON,
+          sections[0].cursorIntOFF,
+          sections[0].endOutput
+        ),
+      INTRO: () =>
+        handleButtonClick(
+          sections[1].bTitle,
+          sections[1].output,
+          sections[1].cursorIntON,
+          sections[1].cursorIntOFF,
+          sections[1].endOutput
+        ),
+      CHECK: () =>
+        handleButtonClick(
+          sections[2].bTitle,
+          sections[2].output,
+          sections[2].cursorIntON,
+          sections[2].cursorIntOFF,
+          sections[2].endOutput
+        ),
+      TIP: () =>
+        handleButtonClick(
+          sections[3].bTitle,
+          sections[3].output,
+          sections[3].cursorIntON,
+          sections[3].cursorIntOFF,
+          sections[3].endOutput
+        ),
+      PROTIP: () =>
+        handleButtonClick(
+          sections[4].bTitle,
+          sections[4].output,
+          sections[4].cursorIntON,
+          sections[4].cursorIntOFF,
+          sections[4].endOutput
+        ),
+      CHALLENGE: () =>
+        handleButtonClick(
+          sections[5].bTitle,
+          sections[5].output,
+          sections[5].cursorIntON,
+          sections[5].cursorIntOFF,
+          sections[5].endOutput
+        ),
+      FLAG: () =>
+        handleButtonClick(
+          sections[6].bTitle,
+          sections[6].output,
+          sections[6].cursorIntON,
+          sections[6].cursorIntOFF,
+          sections[6].endOutput
+        ),
+      TRY: () =>
+        handleButtonClick(
+          sections[7].bTitle,
+          sections[7].output,
+          sections[7].cursorIntON,
+          sections[7].cursorIntOFF,
+          sections[7].endOutput
+        ),
+
+      //code
+      INLINE: () =>
+        handleButtonClick(
+          code[0].bTitle,
+          code[0].output,
+          code[0].cursorIntON,
+          code[0].cursorIntOFF,
+          code[0].endOutput
+        ),
+      CODEBLOCK: () =>
+        handleButtonClick(
+          code[1].bTitle,
+          code[1].output,
+          code[1].cursorIntON,
+          code[1].cursorIntOFF,
+          code[1].endOutput
+        )
+    };
+
+    const handlePreview = event => {
+      if (event) {
+        if (this.state.preview) {
+          this.setState({ preview: false });
+        } else {
+          this.setState({ preview: true });
+        }
       }
     };
 
     return this.props.isSignedIn ? (
-      <div>
-        {!this.state.showPreview ? (
-          <div>
+      this.state.preview ? (
+        <div className="ui grid">
+          <div className="row">
+            <div className="column">
+              <Button style={{}} onClick={() => handlePreview(true)}>
+                <Icon name="eye" />
+              </Button>
+            </div>
+          </div>
+          <div
+            id="MDPreview"
+            style={{ display: "block" }}
+            className="eight wide column"
+          >
+            <MDPreview />
+          </div>
+        </div>
+      ) : (
+        <div className="ui grid">
+          <div className="row">
             <ControlPanel
-              returnPreview={returnPreview}
               handleButtonClick={handleButtonClick}
               nextTitle={NAV_BUTTONS.submit}
               prevValue="/createNewLesson"
               nextValue="/endpage"
-              mySubmitHandler={mySubmitHandler}
+              submitHandler={submitHandler}
+              handlePreview={handlePreview}
+            />
+          </div>
+          <div id="MDTextArea" className="eight wide column">
+            <MDTextArea
+              editorRef={this.editorRef}
+              onInputChange={handleChange}
+              handleButtonClick={handleButtonClick}
+              onTextareaKeyDown={onTextareaKeyDown}
+              onTextareaKeyUp={onTextareaKeyUp}
+              onTextareaMouseDown={onTextareaMouseDown}
+              onTextareaSelect={onTextareaSelect}
+              autoSaveMessage={autoSaveMessage}
+              handlers={handlers}
+              keyMap={keyMap}
+            />
+          </div>
+          {imagePopup}
+          <div id="MDPreview" className="eight wide column">
+            <MDPreview />
+          </div>
+          <div className="row">
+            <PageButtons
+              prevTitle={NAV_BUTTONS.prev}
+              nextTitle={NAV_BUTTONS.submit}
+              prevValue="/createNewLesson"
+              nextValue="/endpage"
+              setPageNumber={null}
+              submitHandler={submitHandler}
               state={this.state}
             />
-            <div className="">
-              <div className="">
-                <MDTextArea
-                  editorRef={this.editorRef}
-                  textValue={this.state.textValue}
-                  onInputChange={handleChange}
-                  handleButtonClick={handleButtonClick}
-                  onTextareaKeyDown={onTextareaKeyDown}
-                  onTextareaKeyUp={onTextareaKeyUp}
-                  onTextareaMouseDown={onTextareaMouseDown}
-                  onTextareaSelect={onTextareaSelect}
-                  autoSaveMessage={autoSaveMessage}
-                />
-              </div>
-              {imagePopup}
-            </div>
-            <footer>
-              <ControlPanel2 handleButtonClick={handleButtonClick} />
-            </footer>
           </div>
-        ) : (
-          <div>
-            <ControlPanelPreview
-              returnPreview={returnPreview}
-              handleButtonClick={handleButtonClick}
-            />
-            <div className="">
-              <div className="">
-                <MDPreview mdValue={this.state.mdValue} />
-              </div>
-            </div>
-          </div>
-        )}
-      </div>
+        </div>
+      )
     ) : (
       <Redirect to="/" />
     );
