+++ src/components/mdTextEditor/editor/editor.css
@@ -1,0 +1,94 @@
+select {
+  cursor: pointer;
+}
+
+/* ControlPanel  */
+
+@media only screen and (max-width: 955px) {
+  #sections {
+    display: none;
+  }
+  #code {
+    display: none;
+  }
+
+  .preview {
+    display: block !important;
+  }
+
+  #saveLoad {
+    display: none;
+  }
+  #MDPreview {
+    display: none;
+  }
+  #MDTextArea {
+    width: 100% !important;
+  }
+  #MDPreview {
+    width: 100% !important;
+  }
+}
+
+.preview {
+  display: none;
+}
+
+/* CSS for ImagePopup component */
+
+.transparent {
+  height: 150%;
+  width: 150%;
+  z-index: 2;
+  background-color: rgba(0, 0, 0, 0.6);
+  position: absolute;
+
+  top: -10%;
+  left: -25%;
+}
+
+.imagePopup {
+  z-index: 3;
+  top: 30%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  position: absolute;
+}
+
+.errorMessage {
+  color: red;
+}
+
+/* CSS for TextArea */
+
+.TextArea {
+  width: 100%;
+  height: 65vh;
+  outline: none;
+  resize: none;
+  border-radius: 5px;
+  padding: 15px;
+  overflow: auto;
+}
+
+/* CSS for TextArea Buttons */
+
+/* CSS for MDPreview */
+
+.MDPreviewSmall {
+  width: 100%;
+  height: 65vh;
+  background-color: white;
+  padding: 15px;
+  overflow: auto;
+}
+
+.MDPreview {
+  width: 100%;
+  height: 65vh;
+  border: 1px solid black;
+  border-radius: 5px;
+  background-color: white;
+  padding: 15px;
+  overflow: auto;
+}
