+++ src/components/mypage/RightBox.js
@@ -1,18 +1,16 @@
-import React from 'react';
-import "./RightBoxStyle.css";
+import React from "react";
+// import "./RightBoxStyle.css";
 
-function RightBox(){
+function RightBox() {
   return (
-    <div className = "RightBox">
-      <div className = "informationTable">
-      <p id="header">Informasjon</p>
+    <div className="RightBox">
+      <div className="informationTable">
+        <p id="header">Informasjon</p>
       </div>
       <p id="text1">Ferdig behandlet</p>
       <p id="text2">Under behandling</p>
       <p id="text3">Jobbes med</p>
     </div>
-
-  )
-
+  );
 }
 export default RightBox;
