+++ src/index.js
@@ -1,9 +1,10 @@
 // import 'bootstrap/dist/css/bootstrap.css';
 import "semantic-ui-css/semantic.min.css";
+import "./index.css";
 import React from "react";
 import ReactDOM from "react-dom";
 import { Provider } from "react-redux";
-import { createStore, applyMiddleware } from "redux";
+import { createStore } from "redux";
 
 import App from "./components/App";
 import reducers from "./reducers";
