+++ src/reducers/index.js
@@ -2,9 +2,11 @@ import { combineReducers } from "redux";
 import mdTextReducer from "./mdTextReducer";
 import parseMDReducer from "./parseMDReducer";
 import authReducer from "./authReducer";
+import formpageReducer from "./formpageReducer";
 
 export default combineReducers({
   mdText: mdTextReducer,
   parseMD: parseMDReducer,
-  auth: authReducer
+  auth: authReducer,
+  formpageValues: formpageReducer
 });
